# OpenAPIDescription.yaml
openapi: 3.0.0
info:
  description: This is a password manager server.
  version: 0.0.1
  title: Password Manager
  termsOfService: ""
  contact:
    email: test@te.st
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
tags:
  - name: user
    description: Creating users, signing in/out
  - name: accounts
    description: Access to accounts stored in the password manager
  - name: maintenance
    description: General information
paths:
  /info:
    get:
      tags:
        - maintenance
      summary: get information about the server for the client
      description: ""
      operationId: serverInformation
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerInformation"
  /user:
    put:
      tags:
        - user
      summary: registration
      description: ""
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInformation"
        description: Registration Object
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccessMessage"
        "405":
          description: Invalid input
      security:
        - csrf: []
  /user/login:
    post:
      tags:
        - user
      summary: login
      description: ""
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogonInformation"
        description: Logon Object
        required: true
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
                example: SESSIONID=1234567890; Path=/; HttpOnly; Secure; SameSite
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/LogonResult"
        "405":
          description: Invalid input
      security:
        - csrf: []
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      description: ""
      operationId: logoutUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccessMessage"
  /user/changepassword:
    post:
      tags:
        - user
      summary: "change user password"
      description: "change password of current user and upload reencrypted accounts"
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePassword"
        description: ChangePassword Object
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccessMessage"
        "403":
          description: Unauthorized
      security:
        - csrf: []
  /user/settings:
    get:
      tags:
        - user
      summary: Returns the client settings of the current user
      description: ""
      operationId: getUserSettings
      responses: 
        "200":
          description: stored client values (encrypted string that contains a JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "403":
          description: Unauthorized
    post:
      tags:
        - user
      summary: "change client settings of current user"
      description: "parameter contains encrypted client settings"
      operationId: setUserSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettings"
        description: Settings Object
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccessMessage"
        "403":
          description: Unauthorized
      security:
        - csrf: []
  /user/history:
    get:
      tags:
        - user
      summary: Returns a history of successful and failed logins
      description: ""
      operationId: getUserHistory
      responses: 
        "200":
          description: Array of logins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserHistory"
        "403":
          description: Unauthorized
  /accounts:
    get:
      tags:
        - accounts
      summary: Returns the accounts stored by the current user
      description: ""
      operationId: getAccounts
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
        "400":
          description: Invalid status value
        "403":
          description: Unauthorized
    put:
      tags:
        - accounts
      summary: Add Account
      description: ""
      operationId: addAccounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAccounts"
        description: Account Values
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
        "400":
          description: Invalid status value
        "403":
          description: Unauthorized
      security:
        - csrf: []
  /accounts/{id}:
    post:
      tags:
        - accounts
      summary: Update a stored account
      description: ""
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the account
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        description: Account Object
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
        "400":
          description: Invalid status value
        "403":
          description: Unauthorized
      security:
        - csrf: []
    delete:
      tags:
        - accounts
      summary: Delete a stored Account
      description: ""
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the account
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
        "400":
          description: Invalid status value
        "403":
          description: Unauthorized
externalDocs:
  description: Repository
  url: http://github.com/zeruniverse/PasswordManager
components:
  securitySchemes:
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
    RegistrationInformation:
      type: object
      allOf:
        - $ref: "#/components/schemas/LogonInformation"
        - $ref: "#/components/schemas/UserInformation"
    LogonInformation:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    LogonResult:
        type: object
        allOf:
          - $ref: "#/components/schemas/GenericSuccessMessage"
          - $ref: "#/components/schemas/LogonSecurityInformation"
    LogonSecurityInformation:
      type: object
      properties:
        lastLogin:
          type: string
          format: date-time
          nullable: true
        failedLogins:
          type: integer
    UserInformation:
      type: object
      properties:
        email:
          type: string
    UserHistory:
      type: array
      items:
        $ref: "#/components/schemas/HistoryItem"
    HistoryItem:
      type: object
      properties:
        UserAgent:
          type: string
        IP:
          type: string
        Time:
          type: string
          format: date-time
        Event:
          type: string
          enum:
            - Login
            - ChangePassword
            - Registration
        EventResult:
          type: string
    UserSettings:
      type: object
      properties:
        encryptedUserSettings:
          type: string
    ChangePassword:
      type: object
      properties:
        newPassword:
          type: string
        accounts:
          $ref: "#/components/schemas/Accounts"
    AccountId:
      allOf:
        - $ref: "#/components/schemas/Index"
        - $ref: "#/components/schemas/Account"
    Account:
      type: object
      properties:
        name:
          type: string
        additional:
          type: string
        password:
          type: string
    Accounts:
      type: array
      items:
        $ref: "#/components/schemas/AccountId"
    NewAccounts:
      type: array
      items:
        $ref: "#/components/schemas/Account"
    Index:
      type: object
      properties:
        index:
          type: integer
    File:
      type: object
      properties:
        index:
          type: integer
        name:
          type: string
        key:
          type: string
    Files:
      type: array
      items:
        $ref: "#/components/schemas/File"
    ServerInformation:
      type: object
      properties:
        csrfToken:
          type: string
        allowRegistration:
          type: boolean
    GenericSuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string